requirements.txt
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.2
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-HTTPAuth==3.3.0
gspread==3.1.0
gunicorn==20.0.4
httplib2==0.15.0
idna==2.8
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.3.1
python-dotenv==0.10.3
requests==2.22.0
rsa==4.0
six==1.12.0
termcolor==1.1.0
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
